a lot of shein improvements including product import and finish product push for both configurable and simple products

+++ OneSila/sales_channels/integrations/shein/constants.py
@@ -208,7 +208,7 @@ SHEIN_INTERNAL_PROPERTY_DEFINITIONS = [
     {
         "code": "weight",
         "name": _("Package weight (g)"),
-        "type": Property.TYPES.INT,
+        "type": Property.TYPES.FLOAT,
         "payload_field": "weight",
     },
 ]
